Moved DateUtility into runtime namespace
