+++ client/Packages/com.beamable/Common/Runtime/Modules/SharedContent/ScheduleDefinition.cs
@@ -1,5 +1,6 @@
 using Beamable.Common.Content.Validation;
 using Beamable.Content;
+using Beamable.Content.Utility;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -33,7 +34,7 @@ namespace Beamable.Common.Content
 		public string description;
 
 		[MustBeDateString]
-		public string activeFrom;
+		public string activeFrom = DateTime.UtcNow.ToString(DateUtility.ISO_FORMAT);
 
 		[MustBeDateString]
 		public OptionalString activeTo = new OptionalString();
