--- client/Packages/com.beamable/Editor/Utility/DateUtility.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 42b7507b803046d2bd2844d25fe536d2
-timeCreated: 1625780383
