+++ client/Packages/com.beamable/Runtime/BeamContext.cs
@@ -15,6 +15,7 @@ using Beamable.Common.Content;
 using Beamable.Common.Dependencies;
 using Beamable.Common.Player;
 using Beamable.Config;
+using Beamable.Content.Utility;
 using Beamable.Coroutines;
 using Beamable.Player;
 using Core.Platform.SDK;
@@ -178,7 +179,7 @@ namespace Beamable
 				}
 
 				var date = DateTime.Parse(value, null, System.Globalization.DateTimeStyles.RoundtripKind);
-				var str = date.ToString("yyyy-MM-ddTHH:mm:ssZ");
+				var str = date.ToString(DateUtility.ISO_FORMAT);
 				_requester.TimeOverride = str;
 				TimeOverrideChanged?.Invoke();
 			}
